#65 Refactor: 패키지를 재배치한다
